Fixed classes issue and moved the data stor factory class into its own file.
